+++ syncany-lib/src/test/java/org/syncany/tests/OtherShortTestSuite.java
@@ -32,7 +32,6 @@ import org.syncany.tests.crypto.CipherSessionTest;
 import org.syncany.tests.crypto.CipherSpecsTest;
 import org.syncany.tests.crypto.CipherUtilTest;
 import org.syncany.tests.crypto.MultiCipherStreamsTest;
-import org.syncany.tests.operations.*;
 import org.syncany.tests.plugins.PluginOptionsTest;
 import org.syncany.tests.plugins.PluginsTest;
 import org.syncany.tests.plugins.TransferSettingsTest;
@@ -66,20 +65,6 @@ import org.syncany.tests.util.SqlRunnerTest;
 		CleanupInterruptedTest.class,
 		TransferSettingsTest.class,
 		PluginOptionsTest.class,
-
-		// Operations
-		CleanupOperationTest.class,
-		ConnectOperationTest.class,
-		FileSystemActionComparatorTest.class,
-		FileSystemActionReconciliatorTest.class,
-		IndexerTest.class,
-		InitOperationTest.class,
-		NotificationListenerTest.class,
-		OperationPerformanceTest.class,
-		PluginOperationTest.class,
-		RecursiveWatcherTest.class,
-		StatusOperationTest.class,
-		SyncUpOperationTest.class,
 
 		// Config
 		CacheTest.class,